首页item数据更新&ListView控件封装&细节处理

+++ app/src/main/java/com/leiyun/appmarket/domain/AppInfo.java
@@ -13,5 +13,5 @@ public class AppInfo {
     public String name;
     public String packageName;
     public long size;
-    public double stars;
+    public float stars;
 }
